requirements.txt
absl-py==2.3.1
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.11.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beartype==0.21.0
beautifulsoup4==4.13.4
bleach==6.2.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
catboost==1.2.8
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
crispy-bootstrap5==2025.6
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==6.0
django-allauth==65.12.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.2.0
fastai==2.8.2
fastcore==1.8.7
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.74.0
gTTS==2.5.4
gunicorn==21.2.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
libclang==18.1.1
lightgbm==4.6.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mlxtend==0.23.4
mpmath==1.3.0
murmurhash==1.0.13
namex==0.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt_einsum==3.4.0
optree==0.17.0
optuna==4.4.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
plum-dispatch==2.5.7
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.29
reportlab==4.4.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
sqlparse==0.5.4
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
triton==3.3.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.3
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==66.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.6.0
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.2
zopfli==0.2.3.post1
